# src/Airslip.Analytics.Api/openapi/analytics.api.2021.11.json
{
  "openapi": "3.0.1",
  "info": {
    "title": "Internal Analytics API",
    "description": "An internal only API used for any UI products",
    "version": "2021.11"
  },
  "servers": [
    {
      "url": "http://localhost:7101"
    }
  ],
  "paths": {
    "/2021.11/banking/transactions": {
      "post": {
        "tags": [
          "Banking"
        ],
        "summary": "Get bank transactions enriched with merchant and director details for a connected business",
        "operationId": "Get Bank Transactions",
        "requestBody": {
          "description": "The bank transaction model within the search query. You can use this to sort or search for any column within the model",
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/OwnedDataSearchModel"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BankTransactionReportModelEntitySearchResponse"
                }
              }
            }
          },
          "400": {
            "description": "Bad Request",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
    "/2021.11/banking/transactions/download": {
      "post": {
        "tags": [
          "Banking"
        ],
        "summary": "Get bank transactions enriched with merchant and director details for a connected business",
        "operationId": "Download Bank Transactions",
        "requestBody": {
          "description": "The bank transaction model within the search query. You can use this to sort or search for any column within the model",
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/OwnedDataSearchModel"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BankTransactionReportModelEntitySearchResponse"
                }
              }
            }
          },
          "400": {
            "description": "Bad Request",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
    "/2021.11/banking/balances/search": {
      "post": {
        "tags": [
          "Banking"
        ],
        "summary": "Search for balances for a connected business",
        "operationId": "Get Account Balances",
        "requestBody": {
          "description": "The account balance model within the search query. You can use this to sort or search for any column within the model",
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/OwnedDataSearchModel"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AccountBalanceReportModelEntitySearchResponse"
                }
              }
            }
          },
          "400": {
            "description": "Bad Request",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
    "/2021.11/commerce/accounts/search": {
      "post": {
        "tags": [
          "Commerce"
        ],
        "summary": "Get all commerce accounts for a connected business",
        "operationId": "Get Commerce Accounts",
        "requestBody": {
          "description": "The commerce transaction model within the search query, you can use this to sort or search for any column within the model",
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/OwnedDataSearchModel"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CommerceProviderModelEntitySearchResponse"
                }
              }
            }
          },
          "400": {
            "description": "Bad Request",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
    "/2021.11/commerce/accounts/download": {
      "post": {
        "tags": [
          "Commerce"
        ],
        "summary": "Get all commerce accounts for a connected business",
        "operationId": "Download Commerce Accounts",
        "requestBody": {
          "description": "The commerce transaction model within the search query. You can use this to sort or search for any column within the model",
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/OwnedDataSearchModel"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DownloadResponse"
                }
              }
            }
          },
          "400": {
            "description": "Bad Request",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
    "/2021.11/commerce/transactions/search": {
      "post": {
        "tags": [
          "Commerce"
        ],
        "summary": "Get and search for commerce transactions for a connected business",
        "operationId": "Get Commerce Transactions",
        "requestBody": {
          "description": "The commerce transaction model within the search query. You can use this to sort or search for any column within the model",
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/OwnedDataSearchModel"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CommerceTransactionReportModelEntitySearchResponse"
                }
              }
            }
          },
          "400": {
            "description": "Bad Request",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
    "/2021.11/commerce/transactions/download": {
      "post": {
        "tags": [
          "Commerce"
        ],
        "summary": "Download commerce transactions for a connected business",
        "operationId": "Download Commerce Transactions",
        "requestBody": {
          "description": "The commerce transaction model within the search query. You can use this to sort or search for any column within the model",
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/OwnedDataSearchModel"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BankTransactionReportModelEntitySearchResponse"
                }
              }
            }
          },
          "400": {
            "description": "Bad Request",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
      "AccountBalanceReportModel": {
        "required": [
          "accountDetail",
          "accountStatus",
          "balance",
          "currencyCode",
          "entityStatus",
          "integrationProviderId",
          "provider",
          "updatedOn"
        ],
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "nullable": true
          },
          "integrationProviderId": {
            "type": "string"
          },
          "accountStatus": {
            "$ref": "#/components/schemas/BankingAccountStatus"
          },
          "sortCode": {
            "type": "string",
            "nullable": true
          },
          "accountNumber": {
            "type": "string",
            "nullable": true
          },
          "currencyCode": {
            "type": "string"
          },
          "balance": {
            "type": "number",
            "format": "double"
          },
          "updatedOn": {
            "type": "string",
            "format": "date-time"
          },
          "provider": {
            "$ref": "#/components/schemas/IntegrationProviderReportModel"
          },
          "accountDetail": {
            "$ref": "#/components/schemas/IntegrationAccountDetailReportModel"
          }
        },
        "additionalProperties": false
      },
      "AccountBalanceReportModelEntitySearchResponse": {
        "required": [
          "paging",
          "results"
        ],
        "type": "object",
        "properties": {
          "results": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/AccountBalanceReportModel"
            }
          },
          "paging": {
            "$ref": "#/components/schemas/EntitySearchPagingModel"
          }
        },
        "additionalProperties": false
      },
...
      "AirslipUserType": {
        "enum": [
          "Unknown",
          "Standard",
          "Merchant",
          "Unregistered",
          "Partner",
          "Administrator",
          "InternalApi"
        ],
        "type": "string"
      },
...
      "AuthenticationState": {
        "enum": [
          "NotAuthenticated",
          "Authenticating",
          "Authenticated"
        ],
        "type": "string"
      },
      "BankTransactionReportModel": {
        "required": [
          "amount",
          "bankTransactionId",
          "capturedDate",
          "description",
          "entityStatus",
          "integration",
          "merchant",
          "merchantTransactionType"
        ],
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "nullable": true
          },
          "bankTransactionId": {
            "type": "string"
          },
          "transactionHash": {
            "type": "string",
            "nullable": true
          },
          "integration": {
            "$ref": "#/components/schemas/IntegrationReportModel"
          },
          "authorisedDate": {
            "type": "integer",
            "format": "int64",
            "nullable": true
          },
          "capturedDate": {
            "type": "integer",
            "format": "int64"
          },
          "amount": {
            "type": "integer",
            "format": "int64"
          },
          "currencyCode": {
            "type": "string",
            "nullable": true
          },
          "description": {
            "type": "string"
          },
          "lastCardDigits": {
            "type": "string",
            "nullable": true
          },
          "isoFamilyCode": {
            "type": "string",
            "nullable": true
          },
          "proprietaryCode": {
            "type": "string",
            "nullable": true
          },
          "reference": {
            "type": "string",
            "nullable": true
          },
          "merchant": {
            "$ref": "#/components/schemas/MerchantResponse"
          },
          "merchantTransactionType": {
            "$ref": "#/components/schemas/BusinessTypes"
          }
        },
        "additionalProperties": false
      },
      "BankTransactionReportModelEntitySearchResponse": {
        "required": [
          "paging",
          "results"
        ],
        "type": "object",
        "properties": {
          "results": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/BankTransactionReportModel"
            }
          },
          "paging": {
            "$ref": "#/components/schemas/EntitySearchPagingModel"
          }
        },
        "additionalProperties": false
      },
      "BankingAccountStatus": {
        "enum": [
          "Active",
          "Expired"
        ],
        "type": "string"
      },
...
      "BusinessTypes": {
        "enum": [
          "Anonymous",
          "Customer",
          "Supplier"
        ],
        "type": "string"
      },
...
      "CommerceProviderModel": {
        "required": [
          "authenticationState",
          "entityStatus",
          "integrationProviderId",
          "name",
          "provider",
          "timeStamp"
        ],
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "nullable": true
          },
          "integrationProviderId": {
            "type": "string"
          },
          "authenticationState": {
            "$ref": "#/components/schemas/AuthenticationState"
          },
          "name": {
            "type": "string"
          },
          "timeStamp": {
            "type": "integer",
            "format": "int64"
          },
          "provider": {
            "$ref": "#/components/schemas/IntegrationProviderReportModel"
          }
        },
        "additionalProperties": false
      },
      "CommerceProviderModelEntitySearchResponse": {
        "required": [
          "paging",
          "results"
        ],
        "type": "object",
        "properties": {
          "results": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/CommerceProviderModel"
            }
          },
          "paging": {
            "$ref": "#/components/schemas/EntitySearchPagingModel"
          }
        },
        "additionalProperties": false
      },
...
      "CommerceTransactionReportModelEntitySearchResponse": {
        "required": [
          "paging",
          "results"
        ],
        "type": "object",
        "properties": {
          "results": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/CommerceTransactionReportModel"
            }
          },
          "paging": {
            "$ref": "#/components/schemas/EntitySearchPagingModel"
          }
        },
        "additionalProperties": false
      },
...
      "DownloadResponse": {
        "required": [
          "fileContent",
          "fileName",
          "mediaType"
        ],
        "type": "object",
        "properties": {
          "fileName": {
            "type": "string"
          },
          "fileContent": {
            "type": "string",
            "format": "byte"
          },
          "mediaType": {
            "type": "string"
          }
        },
        "additionalProperties": false
      },
      "EntitySearchModel": {
        "required": [
          "items",
          "linkOperator"
        ],
        "type": "object",
        "properties": {
          "items": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/SearchFilterModel"
            }
          },
          "linkOperator": {
            "type": "string"
          }
        },
        "additionalProperties": false
      },
      "EntitySearchPagingModel": {
        "required": [
          "next",
          "page"
        ],
        "type": "object",
        "properties": {
          "totalRecords": {
            "type": "integer",
            "format": "int32",
            "nullable": true
          },
          "recordsPerPage": {
            "type": "integer",
            "format": "int32",
            "nullable": true
          },
          "page": {
            "type": "integer",
            "format": "int32"
          },
          "next": {
            "$ref": "#/components/schemas/EntitySearchQueryModel"
          }
        },
        "additionalProperties": false
      },
...
      "EntitySearchSortModel": {
        "required": [
          "field",
          "sort"
        ],
        "type": "object",
        "properties": {
          "field": {
            "type": "string"
          },
          "sort": {
            "$ref": "#/components/schemas/SortOrder"
          }
        },
        "additionalProperties": false
      },
...
      "ErrorResponse": {
        "required": [
          "errorCode",
          "links",
          "metadata"
        ],
        "type": "object",
        "properties": {
          "errorCode": {
            "type": "string"
          },
          "message": {
            "type": "string",
            "nullable": true
          },
          "metadata": {
            "type": "object",
            "additionalProperties": { }
          },
          "links": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/Link"
            }
          }
        },
        "additionalProperties": false
      },
...
      "IntegrationReportModel": {
        "required": [
          "accountDetail",
          "name",
          "provider"
        ],
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "nullable": true
          },
          "name": {
            "type": "string"
          },
          "accountDetail": {
            "$ref": "#/components/schemas/IntegrationAccountDetailReportModel"
          },
          "provider": {
            "$ref": "#/components/schemas/IntegrationProviderReportModel"
          }
        },
        "additionalProperties": false
      },
...
      "Link": {
        "required": [
          "href",
          "method",
          "rel"
        ],
        "type": "object",
        "properties": {
          "href": {
            "type": "string"
          },
          "rel": {
            "type": "string"
          },
          "method": {
            "type": "string"
          }
        },
        "additionalProperties": false
      },
...
      "OwnedDataSearchModel": {
        "required": [
          "ownerAirslipUserType",
          "ownerEntityId",
          "page",
          "recordsPerPage",
          "search",
          "sort"
        ],
        "type": "object",
        "properties": {
          "ownerEntityId": {
            "type": "string"
          },
          "ownerAirslipUserType": {
            "$ref": "#/components/schemas/AirslipUserType"
          },
          "page": {
            "type": "integer",
            "format": "int32"
          },
          "recordsPerPage": {
            "type": "integer",
            "format": "int32"
          },
          "sort": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/EntitySearchSortModel"
            }
          },
          "search": {
            "$ref": "#/components/schemas/EntitySearchModel"
          }
        },
        "additionalProperties": false
      },
...
      "SearchFilterModel": {
        "required": [
          "columnField",
          "operatorValue"
        ],
        "type": "object",
        "properties": {
          "columnField": {
            "type": "string"
          },
          "value": {
            "nullable": true
          },
          "operatorValue": {
            "type": "string"
          }
        },
        "additionalProperties": false
      },
...
      "SortOrder": {
        "enum": [
          "Asc",
          "Desc"
        ],
        "type": "string"
      },
...
    },
    "securitySchemes": {
      "Bearer": {
        "type": "apiKey",
        "description": "e.g Bearer Api_Key.\n\nRequests to the Airslip API are authenticated using the applications `Api_Key`. You can view and manage your credentials in the Airslip Dashboard.\n\nAn Api_Key provides connectivity to all authenticated Airslip API endpoints, so it is important to keep these credentials secure. Do not share your Api_Key in publicly accessible areas such as GitHub, client-side code, or easily accessible configuration settings.\n\nAuthentication is performed using Bearer Authentication. Your Api_Key should be sent as the token.\n\nAll requests should be made via HTTPS.",
        "name": "Authorization",
        "in": "header"
      }
    }
  },
  "security": [
    {
      "Bearer": [ ]
    }
  ],
...
}